requirements.txt
fastapi
uvicorn
pydantic
python-jose
sqlalchemy
databases
python-decouple
psycopg2-binary
requests
